Fix: Integer conversion warning

+++ test/src/test_rs_reed_solomon_codes.cpp
@@ -24,7 +24,7 @@ TEST(TestReedSolomonCodes, test_construct)
 
     // The random generator
     boost::random::mt19937 random_generator;
-    random_generator.seed(time(0));
+    random_generator.seed(static_cast<uint32_t>(time(0)));
 
 
     {
@@ -34,9 +34,9 @@ TEST(TestReedSolomonCodes, test_construct)
         kodo::rs8_decoder::factory fdec(255, 1400);
         kodo::rs8_decoder::pointer pdec = fdec.build(128, 1400);
 
-        
+
     }
-    
+
 }
 
 
@@ -49,15 +49,15 @@ TEST(TestReedSolomonCodes, test_encode_decode)
 
     // The random generator
     boost::random::mt19937 random_generator;
-    random_generator.seed(time(0));
+    random_generator.seed(static_cast<uint32_t>(time(0)));
 
 
     {
         uniform_int choose_symbols
-            = uniform_int(1, fifi::binary8::order - 1); 
+            = uniform_int(1, fifi::binary8::order - 1);
 
         uint32_t symbols = choose_symbols(random_generator);
-        
+
         kodo::rs8_encoder::factory fenc(255, 1400);
         kodo::rs8_encoder::pointer encoder = fenc.build(symbols, 1400);
 
@@ -72,7 +72,7 @@ TEST(TestReedSolomonCodes, test_encode_decode)
 
         kodo::random_uniform<uint8_t> fill_data;
         fill_data.generate(&data_in[0], data_in.size());
-        
+
         kodo::set_symbols(kodo::storage(data_in), encoder);
 
         uint32_t symbol_count = 0;
@@ -80,8 +80,8 @@ TEST(TestReedSolomonCodes, test_encode_decode)
         {
             encoder->encode( &payload[0] );
 
-             
-            
+
+
             decoder->decode( &payload[0] );
             ++symbol_count;
         }
@@ -89,13 +89,13 @@ TEST(TestReedSolomonCodes, test_encode_decode)
         // A reed solomon code should be able to decode
         // with exactly k symbols
         EXPECT_EQ(symbol_count, symbols);
-    
+
         std::vector<uint8_t> data_out(decoder->block_size(), '\0');
-        kodo::copy_symbols(kodo::storage(data_out), decoder); 
-    
+        kodo::copy_symbols(kodo::storage(data_out), decoder);
+
         EXPECT_TRUE(std::equal(data_out.begin(), data_out.end(), data_in.begin()));
 
     }
-    
+
 }
